Implement hacky discord rpc. @TODO get rid of mediaDetection call in main.ts.
